feat: qr code scanner component
