+++ pages/settings/wallets/WalletConnection.tsx
@@ -3,10 +3,9 @@ import React from "react";
 import * as Clipboard from "expo-clipboard";
 import { nwc } from "@getalby/sdk";
 import {
-  Camera as CameraIcon,
   ClipboardPaste,
   Hotel,
-  X,
+  X
 } from "~/components/Icons";
 import { useAppStore } from "lib/state/appStore";
 import { Camera } from "expo-camera/legacy"; // TODO: check if Android camera detach bug is fixed and update camera
@@ -18,8 +17,8 @@ import Toast from "react-native-toast-message";
 import { errorToast } from "~/lib/errorToast";
 import { Nip47Capability } from "@getalby/sdk/dist/NWCClient";
 import Loading from "~/components/Loading";
-import { FocusableCamera } from "~/components/FocusableCamera";
 import { DemoWallets } from "./DemoWallets";
+import QRCodeScanner from "~/components/QRCodeScanner";
 
 export function WalletConnection() {
   const hasConnection = useAppStore((store) => !!store.nwcClient);
@@ -38,13 +37,7 @@ export function WalletConnection() {
   }
 
   const handleScanned = (data: string) => {
-    setScanning((current) => {
-      if (current === true) {
-        // console.log(`Bar code with data ${data} has been scanned!`);
-        connect(data);
-      }
-      return false;
-    });
+    return connect(data);
   };
 
   React.useEffect(() => {
@@ -178,19 +171,7 @@ export function WalletConnection() {
           )}
           {!isConnecting && (
             <>
-              {isScanning && <FocusableCamera onScanned={handleScanned} />}
-              {!isScanning && (
-                <>
-                  <View className="flex-1 h-full flex flex-col items-center justify-center gap-5">
-                    <CameraIcon className="text-foreground" size={64} />
-                    <Text className="text-2xl text-foreground">Camera Permissions Needed</Text>
-                    <Button onPress={scan}>
-                      <Text>Grant Permissions</Text>
-                    </Button>
-                  </View>
-                </>
-              )}
-
+              <QRCodeScanner onScanned={handleScanned} />
               <View className="flex flex-row items-stretch justify-center gap-4 p-6">
                 <Button
                   variant="secondary"
